Projects/marketing-dashboard/loops_webhook/requirements.txt
Flask==3.0.0
gspread==6.1.2
google-auth==2.40.1
